fix dtype

+++ policy/learner.py
@@ -76,7 +76,7 @@ class Learner:
         self._train_step = tf.Variable(
             0,
             trainable=False,
-            dtype=tf.uint64,
+            dtype=tf.int32,
             aggregation=tf.VariableAggregation.ONLY_FIRST_REPLICA,
             shape=()
         )
